feat: complete CVService cleanup and eliminate legacy code
BREAKING CHANGE: Remove cv_service.py entirely and move all functions to appropriate services

## Major Changes:
- 🗑️ DELETE: backend/app/services/cv_service.py (340+ lines of legacy code)
- ➕ ADD: backend/app/utils/timing_utils.py (timing utilities)
- 🔄 MOVE: Video utilities to video_service.py
- 🔄 MOVE: Contact detection functions to ball_contact_service.py
- 🔄 MOVE: Timing utilities to utils/timing_utils.py

## Function Relocations:
### Video Service (video_service.py):
- extract_frames() - Frame extraction with frame skipping logic
- get_video_metadata() - Video metadata extraction

### Ball Contact Service (ball_contact_service.py):
- detect_ball_contact() - Basic ball contact detection
- detect_ball_contact_with_rackets() - Advanced racket-based detection
- _calculate_racket_head_position() - Racket head position calculation
- _calculate_ball_trajectory_change() - Trajectory analysis

### Utils (utils/timing_utils.py):
- log_timing() - Success timing logging
- log_timing_error() - Error timing logging

## Database Changes:
- Remove legacy analysis_id foreign key from VideoAnnotation model
- Add migration to remove analysis_id column from video_annotations table
- Fix all test failures related to missing analyses table

## Test Updates:
- Update all test imports to use new function locations
- Fix test assertions to match new function return formats
- All 19 tests now passing ✅

## Documentation Updates:
- Update README.md to reflect new service structure
- Mark analysis_backend_refactoring_plan.md as complete
- Remove references to cv_service.py

## Code Quality:
- All linting issues resolved ✅
- No legacy dependencies remaining
- Clean, modular architecture with proper separation of concerns

## Impact:
- Removed 340+ lines of legacy code from CVService
- Total reduction: 2000+ lines of legacy code across entire refactoring
- Backend now runs entirely on modular services
- No breaking changes to production code (CVService was unused)

This completes the analysis backend refactoring project.
